requirements.txt
camel-ai[all]==0.2.16
python-dotenv
notebook
pandas
langchain
langchain-openai
tiktoken
neo4j
unstructured